pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "libucks"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "sentence-transformers>=3.0",
    "numpy>=1.26",
    "anthropic>=0.40",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "tree-sitter>=0.22",
    "gitpython>=3.1",
    "scipy>=1.13",
    "scikit-learn>=1.5",
    "watchdog>=4.0",
    "unidiff>=0.7",
    "click>=8.1",
    "rich>=13.0",
    "mcp>=1.0",
    "structlog>=24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-timeout>=2.3",
    "respx>=0.21",
    "jsonschema>=4.22",
]

[project.scripts]
libucks = "main:cli"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "slow: marks tests that load the embedding model (deselect with '-m not slow')",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["libucks*"]
